configuration de la datasource dans wildwfly le drivers postgresql ne depend pas encore de jarkarta, il depend de javax

+++ calipso-dto/src/main/java/com/jdevhub/calipso/dto/AccountDto.java
@@ -5,6 +5,6 @@ import java.time.LocalDateTime;
 
 
 
-public record AccountDto(Long id, String number, BigDecimal balance, LocalDateTime creationDate, CustomerDto customer) {
+public record AccountDto(Integer id, String number, BigDecimal balance, LocalDateTime creationDate, CustomerDto customer) {
 
 }
